+++ ql/src/java/org/apache/hadoop/hive/ql/ddl/table/create/show/ShowCreateTableOperation.java
@@ -47,6 +47,7 @@ import org.apache.hadoop.hive.ql.ddl.DDLOperation;
 import org.apache.hadoop.hive.ql.ddl.DDLOperationContext;
 import org.apache.hadoop.hive.ql.ddl.ShowUtils;
 import org.apache.hadoop.hive.ql.ddl.table.create.CreateTableOperation;
+import org.apache.hadoop.hive.ql.exec.DDLPlanUtils;
 import org.apache.hadoop.hive.ql.metadata.HiveException;
 import org.apache.hadoop.hive.ql.metadata.Table;
 import org.apache.hadoop.hive.ql.util.DirectionUtils;
@@ -90,8 +91,9 @@ public class ShowCreateTableOperation extends DDLOperation<ShowCreateTableDesc> 
     // get the create table statement for the table and populate the output
     try (DataOutputStream outStream = ShowUtils.getOutputStream(new Path(desc.getResFile()), context)) {
       Table table = context.getDb().getTable(desc.getDatabaseName(), desc.getTableName());
-      String command = table.isView() ? getCreateViewCommand(table, desc.isRelative())
-          : getCreateTableCommand(table, desc.isRelative());
+      DDLPlanUtils ddlObj = new DDLPlanUtils();
+      String command = table.isView() ? ddlObj.getCreateViewCommand(table, desc.isRelative())
+          : ddlObj.getCreateTableCommand(table, desc.isRelative());
 
       outStream.write(command.getBytes(StandardCharsets.UTF_8));
       return 0;
@@ -101,270 +103,5 @@ public class ShowCreateTableOperation extends DDLOperation<ShowCreateTableDesc> 
     } catch (Exception e) {
       throw new HiveException(e);
     }
-  }
-
-  private static final String CREATE_VIEW_TEMPLATE =
-      "CREATE VIEW <if(" + DATABASE_NAME + ")>`<" + DATABASE_NAME + ">`.<endif>`<" + TABLE_NAME + ">`<" + PARTITIONS + "> AS <SQL>";
-
-  private String getCreateViewCommand(Table table, boolean isRelative) {
-    ST command = new ST(CREATE_VIEW_TEMPLATE);
-
-    if (!isRelative) {
-      command.add(DATABASE_NAME, table.getDbName());
-    }
-    command.add(TABLE_NAME, table.getTableName());
-    command.add(PARTITIONS, getPartitionsForView(table));
-    command.add("SQL", table.getViewExpandedText());
-
-    return command.render();
-  }
-
-  private static final String CREATE_TABLE_TEMPLATE =
-      "CREATE <" + TEMPORARY + "><" + EXTERNAL + ">TABLE <if(" + DATABASE_NAME + ")>`<" + DATABASE_NAME + ">`.<endif>"
-          + "`<" + TABLE_NAME + ">`(\n" +
-      "<" + LIST_COLUMNS + ">)\n" +
-      "<" + COMMENT + ">\n" +
-      "<" + PARTITIONS + ">\n" +
-      "<" + BUCKETS + ">\n" +
-      "<" + SKEWED + ">\n" +
-      "<" + ROW_FORMAT + ">\n" +
-      "<" + LOCATION_BLOCK + ">" +
-      "TBLPROPERTIES (\n" +
-      "<" + PROPERTIES + ">)\n";
-
-  private String getCreateTableCommand(Table table, boolean isRelative) {
-    ST command = new ST(CREATE_TABLE_TEMPLATE);
-
-    if (!isRelative) {
-      command.add(DATABASE_NAME, table.getDbName());
-    }
-    command.add(TABLE_NAME, table.getTableName());
-    command.add(TEMPORARY, getTemporary(table));
-    command.add(EXTERNAL, getExternal(table));
-    command.add(LIST_COLUMNS, getColumns(table));
-    command.add(COMMENT, getComment(table));
-    command.add(PARTITIONS, getPartitions(table));
-    command.add(BUCKETS, getBuckets(table));
-    command.add(SKEWED, getSkewed(table));
-    command.add(ROW_FORMAT, getRowFormat(table));
-    command.add(LOCATION_BLOCK, getLocationBlock(table));
-    command.add(PROPERTIES, getProperties(table));
-
-    return command.render();
-  }
-
-  private String getTemporary(Table table) {
-    return table.isTemporary() ? "TEMPORARY " : "";
-  }
-
-  private String getExternal(Table table) {
-    return table.getTableType() == TableType.EXTERNAL_TABLE ? "EXTERNAL " : "";
-  }
-
-  private String getColumns(Table table) {
-    List<String> columnDescs = new ArrayList<String>();
-    for (FieldSchema column : table.getCols()) {
-      String columnType = formatType(TypeInfoUtils.getTypeInfoFromTypeString(column.getType()));
-      String columnDesc = "  `" + column.getName() + "` " + columnType;
-      if (column.getComment() != null) {
-        columnDesc += " COMMENT '" + HiveStringUtils.escapeHiveCommand(column.getComment()) + "'";
-      }
-      columnDescs.add(columnDesc);
-    }
-    return StringUtils.join(columnDescs, ", \n");
-  }
-
-  /** Struct fields are identifiers, need to be put between ``. */
-  private String formatType(TypeInfo typeInfo) {
-    switch (typeInfo.getCategory()) {
-    case PRIMITIVE:
-      return typeInfo.getTypeName();
-    case STRUCT:
-      StringBuilder structFormattedType = new StringBuilder();
-
-      StructTypeInfo structTypeInfo = (StructTypeInfo)typeInfo;
-      for (int i = 0; i < structTypeInfo.getAllStructFieldNames().size(); i++) {
-        if (structFormattedType.length() != 0) {
-          structFormattedType.append(", ");
-        }
-
-        String structElementName = structTypeInfo.getAllStructFieldNames().get(i);
-        String structElementType = formatType(structTypeInfo.getAllStructFieldTypeInfos().get(i));
-
-        structFormattedType.append("`" + structElementName + "`:" + structElementType);
-      }
-      return "struct<" + structFormattedType.toString() + ">";
-    case LIST:
-      ListTypeInfo listTypeInfo = (ListTypeInfo)typeInfo;
-      String elementType = formatType(listTypeInfo.getListElementTypeInfo());
-      return "array<" + elementType + ">";
-    case MAP:
-      MapTypeInfo mapTypeInfo = (MapTypeInfo)typeInfo;
-      String keyTypeInfo = mapTypeInfo.getMapKeyTypeInfo().getTypeName();
-      String valueTypeInfo = formatType(mapTypeInfo.getMapValueTypeInfo());
-      return "map<" + keyTypeInfo + "," + valueTypeInfo + ">";
-    case UNION:
-      StringBuilder unionFormattedType = new StringBuilder();
-
-      UnionTypeInfo unionTypeInfo = (UnionTypeInfo)typeInfo;
-      for (TypeInfo unionElementTypeInfo : unionTypeInfo.getAllUnionObjectTypeInfos()) {
-        if (unionFormattedType.length() != 0) {
-          unionFormattedType.append(", ");
-        }
-
-        String unionElementType = formatType(unionElementTypeInfo);
-        unionFormattedType.append(unionElementType);
-      }
-      return "uniontype<" + unionFormattedType.toString() + ">";
-    default:
-      throw new RuntimeException("Unknown type: " + typeInfo.getCategory());
-    }
-  }
-
-  private String getComment(Table table) {
-    String comment = table.getProperty("comment");
-    return (comment != null) ? "COMMENT '" + HiveStringUtils.escapeHiveCommand(comment) + "'" : "";
-  }
-
-  private String getPartitionsForView(Table table) {
-    List<FieldSchema> partitionKeys = table.getPartCols();
-    if (partitionKeys.isEmpty()) {
-      return "";
-    }
-    List<String> partitionCols = new ArrayList<String>();
-    for(String col:table.getPartColNames()) {
-      partitionCols.add('`' + col + '`');
-    }
-    return " PARTITIONED ON (" + StringUtils.join(partitionCols, ", ") + ")";
-  }
-
-  private String getPartitions(Table table) {
-    List<FieldSchema> partitionKeys = table.getPartitionKeys();
-    if (partitionKeys.isEmpty()) {
-      return "";
-    }
-
-    List<String> partitionDescs = new ArrayList<String>();
-    for (FieldSchema partitionKey : partitionKeys) {
-      String partitionDesc = "  `" + partitionKey.getName() + "` " + partitionKey.getType();
-      if (partitionKey.getComment() != null) {
-        partitionDesc += " COMMENT '" + HiveStringUtils.escapeHiveCommand(partitionKey.getComment()) + "'";
-      }
-      partitionDescs.add(partitionDesc);
-    }
-    return "PARTITIONED BY ( \n" + StringUtils.join(partitionDescs, ", \n") + ")";
-  }
-
-  private String getBuckets(Table table) {
-    List<String> bucketCols = table.getBucketCols();
-    if (bucketCols.isEmpty()) {
-      return "";
-    }
-
-    String buckets = "CLUSTERED BY ( \n  " + StringUtils.join(bucketCols, ", \n  ") + ") \n";
-
-    List<Order> sortColumns = table.getSortCols();
-    if (!sortColumns.isEmpty()) {
-      List<String> sortKeys = new ArrayList<String>();
-      for (Order sortColumn : sortColumns) {
-        String sortKeyDesc = "  " + sortColumn.getCol() + " " + DirectionUtils.codeToText(sortColumn.getOrder());
-        sortKeys.add(sortKeyDesc);
-      }
-      buckets += "SORTED BY ( \n" + StringUtils.join(sortKeys, ", \n") + ") \n";
-    }
-
-    buckets += "INTO " + table.getNumBuckets() + " BUCKETS";
-    return buckets;
-  }
-
-  private String getSkewed(Table table) {
-    SkewedInfo skewedInfo = table.getSkewedInfo();
-    if (skewedInfo == null || skewedInfo.getSkewedColNames().isEmpty()) {
-      return "";
-    }
-
-    List<String> columnValuesList = new ArrayList<String>();
-    for (List<String> columnValues : skewedInfo.getSkewedColValues()) {
-      columnValuesList.add("('" + StringUtils.join(columnValues, "','") + "')");
-    }
-
-    String skewed =
-        "SKEWED BY (" + StringUtils.join(skewedInfo.getSkewedColNames(), ",") + ")\n" +
-        "  ON (" + StringUtils.join(columnValuesList, ",") + ")";
-    if (table.isStoredAsSubDirectories()) {
-      skewed += "\n  STORED AS DIRECTORIES";
-    }
-    return skewed;
-  }
-
-  private String getRowFormat(Table table) {
-    StringBuilder rowFormat = new StringBuilder();
-
-    StorageDescriptor sd = table.getTTable().getSd();
-    SerDeInfo serdeInfo = sd.getSerdeInfo();
-
-    rowFormat
-      .append("ROW FORMAT SERDE \n")
-      .append("  '" + HiveStringUtils.escapeHiveCommand(serdeInfo.getSerializationLib()) + "' \n");
-
-    Map<String, String> serdeParams = serdeInfo.getParameters();
-    if (table.getStorageHandler() == null) {
-      // If serialization.format property has the default value, it will not to be included in SERDE properties
-      if (Warehouse.DEFAULT_SERIALIZATION_FORMAT.equals(serdeParams.get(serdeConstants.SERIALIZATION_FORMAT))) {
-        serdeParams.remove(serdeConstants.SERIALIZATION_FORMAT);
-      }
-      if (!serdeParams.isEmpty()) {
-        appendSerdeParams(rowFormat, serdeParams);
-        rowFormat.append(" \n");
-      }
-      rowFormat
-        .append("STORED AS INPUTFORMAT \n  '" + HiveStringUtils.escapeHiveCommand(sd.getInputFormat()) + "' \n")
-        .append("OUTPUTFORMAT \n  '" + HiveStringUtils.escapeHiveCommand(sd.getOutputFormat()) + "'");
-    } else {
-      String metaTableStorage = table.getParameters().get(META_TABLE_STORAGE);
-      rowFormat.append("STORED BY \n  '" + HiveStringUtils.escapeHiveCommand(metaTableStorage) + "' \n");
-      if (!serdeParams.isEmpty()) {
-        appendSerdeParams(rowFormat, serdeInfo.getParameters());
-      }
-    }
-
-    return rowFormat.toString();
-  }
-
-  public static void appendSerdeParams(StringBuilder builder, Map<String, String> serdeParams) {
-    SortedMap<String, String> sortedSerdeParams = new TreeMap<String, String>(serdeParams);
-    List<String> serdeCols = new ArrayList<String>();
-    for (Entry<String, String> entry : sortedSerdeParams.entrySet()) {
-      serdeCols.add("  '" + entry.getKey() + "'='" +
-          HiveStringUtils.escapeUnicode(HiveStringUtils.escapeHiveCommand(entry.getValue())) + "'");
-    }
-
-    builder
-      .append("WITH SERDEPROPERTIES ( \n")
-      .append(StringUtils.join(serdeCols, ", \n"))
-      .append(')');
-  }
-
-  private static final String CREATE_TABLE_TEMPLATE_LOCATION =
-      "LOCATION\n" +
-      "<" + LOCATION + ">\n";
-
-  private String getLocationBlock(Table table) {
-    if (!CreateTableOperation.doesTableNeedLocation(table)) {
-      return "";
-    }
-
-    ST locationBlock = new ST(CREATE_TABLE_TEMPLATE_LOCATION);
-    StorageDescriptor sd = table.getTTable().getSd();
-    locationBlock.add(LOCATION, "  '" + HiveStringUtils.escapeHiveCommand(sd.getLocation()) + "'");
-    return locationBlock.render();
-  }
-
-  private static final Set<String> PROPERTIES_TO_IGNORE_AT_TBLPROPERTIES = Sets.union(
-      ImmutableSet.<String>of("TEMPORARY", "EXTERNAL", "comment", "SORTBUCKETCOLSPREFIX", META_TABLE_STORAGE, TABLE_IS_CTAS),
-      new HashSet<String>(StatsSetupConst.TABLE_PARAMS_STATS_KEYS));
-
-  private String getProperties(Table table) {
-    return ShowUtils.propertiesToString(table.getParameters(), PROPERTIES_TO_IGNORE_AT_TBLPROPERTIES);
   }
 }
